Revert "feat: add disabled state for the mcp servers page (#894)"
This reverts commit ab213ea48877bb5bfbb6a124a84cf5885ec2db12.
