+++ renderer/src/features/clients/components/manage-clients-button.tsx
@@ -1,6 +1,13 @@
+import { usePrompt } from '@/common/hooks/use-prompt'
+import type { UseFormReturn } from 'react-hook-form'
+import { Label } from '@/common/components/ui/label'
 import { Button } from '@/common/components/ui/button'
+import { Switch } from '@/common/components/ui/switch'
 import { Code } from 'lucide-react'
-import { useManageClientsDialog } from '../hooks/use-manage-clients-dialog'
+import { z } from 'zod/v4'
+import { zodV4Resolver } from '@/common/lib/zod-v4-resolver'
+import { useManageClients } from '../hooks/use-manage-clients'
+import { useToastMutation } from '@/common/hooks/use-toast-mutation'
 
 interface ManageClientsButtonProps {
   groupName: string
@@ -19,12 +26,81 @@ export function ManageClientsButton({
   variant = 'outline',
   className,
 }: ManageClientsButtonProps) {
-  const { openDialog } = useManageClientsDialog(groupName)
+  const promptForm = usePrompt()
+
+  const {
+    installedClients,
+    defaultValues,
+    reconcileGroupClients,
+    getClientFieldName,
+  } = useManageClients(groupName)
+
+  const { mutateAsync: saveClients } = useToastMutation({
+    mutationFn: reconcileGroupClients,
+    loadingMsg: 'Saving client settings...',
+    successMsg: 'Client settings saved',
+    errorMsg: 'Failed to save client settings',
+  })
+
+  const handleManageClients = async () => {
+    const formSchema = z.object(
+      installedClients.reduce(
+        (acc, client) => {
+          const fieldName = getClientFieldName(client.client_type!)
+          acc[fieldName] = z.boolean()
+          return acc
+        },
+        {} as Record<string, z.ZodBoolean>
+      )
+    )
+
+    const result = await promptForm({
+      title: 'Manage Clients',
+      defaultValues,
+      resolver: zodV4Resolver(formSchema),
+      fields: (form: UseFormReturn<Record<string, boolean>>) => (
+        <div className="rounded-xl border">
+          {installedClients.map((client) => {
+            const fieldName = getClientFieldName(client.client_type!)
+            const displayName = client.client_type!
+
+            return (
+              <div
+                key={client.client_type}
+                className="flex items-start gap-2 border-b p-4 align-middle last:border-b-0"
+              >
+                <Switch
+                  id={fieldName}
+                  checked={form.watch(fieldName) as boolean}
+                  onCheckedChange={(checked) => {
+                    form.setValue(fieldName, checked)
+                    form.trigger(fieldName)
+                  }}
+                />
+
+                <Label htmlFor={fieldName} className="text-sm font-medium">
+                  {displayName}
+                </Label>
+              </div>
+            )
+          })}
+        </div>
+      ),
+      buttons: {
+        confirm: 'Save',
+        cancel: 'Cancel',
+      },
+    })
+
+    if (result) {
+      await saveClients(result)
+    }
+  }
 
   return (
     <Button
       variant={variant}
-      onClick={() => openDialog()}
+      onClick={handleManageClients}
       className={className}
     >
       <Code className="mr-2 h-4 w-4" />
