+++ renderer/src/features/mcp-servers/components/card-mcp-server/index.tsx
@@ -94,7 +94,6 @@ export function CardMcpServer({
   url,
   remote,
   transport,
-  isInDisabledGroup,
 }: {
   name: string
   status: CoreWorkload['status']
@@ -102,7 +101,6 @@ export function CardMcpServer({
   remote?: CoreWorkload['remote']
   url: string
   transport: CoreWorkload['transport_type']
-  isInDisabledGroup?: boolean
 }) {
   const isRemoteMcpEnabled = useFeatureFlag(featureFlagKeys.REMOTE_MCP)
   const nameRef = useRef<HTMLElement | null>(null)
@@ -154,7 +152,6 @@ export function CardMcpServer({
         'transition-all duration-300 ease-in-out',
         isNewServer ? 'ring-2' : undefined,
         isDeleting ? 'pointer-events-none opacity-50' : undefined,
-        isInDisabledGroup ? 'opacity-50 grayscale' : undefined,
         (isTransitioning || hadRecentStatusChange) && 'animate-diagonal-ring',
         isStopped && 'bg-card/65'
       )}
