+++ renderer/src/features/mcp-servers/components/grid-cards-mcp-server.tsx
@@ -5,10 +5,8 @@ import { InputSearch } from '@/common/components/ui/input-search'
 
 export function GridCardsMcpServers({
   mcpServers,
-  isInDisabledGroup,
 }: {
   mcpServers: CoreWorkload[]
-  isInDisabledGroup?: boolean
 }) {
   const [filters, setFilters] = useState({
     text: '',
@@ -59,7 +57,6 @@ export function GridCardsMcpServers({
               statusContext={mcpServer.status_context}
               url={mcpServer.url ?? ''}
               transport={mcpServer.transport_type}
-              isInDisabledGroup={isInDisabledGroup}
             />
           ) : null
         )}
