+++ renderer/src/routes/group.$groupName.tsx
@@ -14,9 +14,7 @@ import { RefreshButton } from '@/common/components/refresh-button'
 import { DropdownMenuRunMcpServer } from '@/features/mcp-servers/components/dropdown-menu-run-mcp-server'
 import { WrapperDialogFormMcp } from '@/features/mcp-servers/components/wrapper-dialog-mcp'
 import { ManageClientsButton } from '@/features/clients/components/manage-clients-button'
-import { EnableGroupButton } from '@/features/clients/components/enable-group-button'
 import { GroupActionsDropdown } from '@/features/mcp-servers/components/group-actions-dropdown'
-import { useGroups } from '@/features/mcp-servers/hooks/use-groups'
 
 export const Route = createFileRoute('/group/$groupName')({
   loader: ({ context: { queryClient }, params: { groupName } }) =>
@@ -45,14 +43,6 @@ function GroupRoute() {
   })
 
   const workloads = data?.workloads ?? []
-  const { data: groupsData } = useGroups()
-  const currentGroup = (groupsData?.groups ?? []).find(
-    (g) => (g.name ?? '').toLowerCase() === (groupName ?? '').toLowerCase()
-  )
-  const isInDisabledGroup = !(
-    currentGroup?.registered_clients &&
-    currentGroup.registered_clients.length > 0
-  )
   const filteredWorkloads = workloads
   const [serverDialogOpen, setServerDialogOpen] = useState<{
     local: boolean
@@ -97,16 +87,10 @@ function GroupRoute() {
               {workloads.length > 0 && (
                 <>
                   <RefreshButton refresh={refetch} />
-                  {!isInDisabledGroup ? (
-                    <DropdownMenuRunMcpServer
-                      openRunCommandDialog={setServerDialogOpen}
-                    />
-                  ) : null}
-                  {!isInDisabledGroup ? (
-                    <ManageClientsButton groupName={groupName} />
-                  ) : (
-                    <EnableGroupButton groupName={groupName} />
-                  )}
+                  <DropdownMenuRunMcpServer
+                    openRunCommandDialog={setServerDialogOpen}
+                  />
+                  <ManageClientsButton groupName={groupName} />
                 </>
               )}
               {showSidebar ? (
@@ -129,19 +113,14 @@ function GroupRoute() {
             body="Add a server manually or browse the MCP Server registry"
             illustration={IllustrationNoConnection}
           >
-            {!isInDisabledGroup ? (
-              <div className="my-6">
-                <DropdownMenuRunMcpServer
-                  openRunCommandDialog={setServerDialogOpen}
-                />
-              </div>
-            ) : null}
+            <div className="my-6">
+              <DropdownMenuRunMcpServer
+                openRunCommandDialog={setServerDialogOpen}
+              />
+            </div>
           </EmptyState>
         ) : (
-          <GridCardsMcpServers
-            mcpServers={filteredWorkloads}
-            isInDisabledGroup={isInDisabledGroup}
-          />
+          <GridCardsMcpServers mcpServers={filteredWorkloads} />
         )}
       </div>
     </div>
